+++ docs/main/changelog.rst
@@ -7,6 +7,10 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Add tests for ``Selector`` provider.
+
 4.17.0
 ------
 - Add ``FastAPI`` + ``SQLAlchemy`` example.